firewalld: Add helper to instal WFD firewall zone
For WFD P2P operation we need to ensure that the sink can connect to the
RTSP port. To do this, add code that dynamically creates the relevant
zone if necessary.

Note that this is just designed as fallback. When installed through a
package manager, then the zone should be installed that way (support for
this will be added by a later commit). That way the user will not be
prompted for a password unless e.g. using the program through a Flatpak.