meta-provider: Fix sink removal in case matches are gone
It may (currently) happen that a sink is reporting incorrect matches at
the time it is removed. In that case, we will not find it in the hash
table, causing us to not find the corresponding meta sink.

Fix this by just doing a linear search for the meta sink that contains
the sink.

This probably means the fix from 23164b58f4d5 (p2p-sink: Guard against
returning NULL as possible match string) is not needed anymore. i.e.
that commit, while fixing a problem, did not address the original crash.

Fixes: #160