wfd-p2p-provider: Do not stop find for all state changes
Previously, we would explicitly start/stop the find operation when the
underlying NMDevice state changed. However, this causes frequent stop
operations to be emitted, which appears to interfer with connection
attempts at a lower level.

Change it to only emit a stop find if the find restart logic is running.

This was changed in commit e870aad3 ("wfd-p2p-provider: Only start
discovery when device available").

Fixes: #257