screencast-portal: embed cursor in stream, if available
Confirm that the embedded cursor mode is available by checking the value
of the property "AvailableCursorModes". For that property to be loaded,
replace the flag G_DBUS_PROXY_FLAGS_DO_NOT_LOAD_PROPERTIES with the flag
G_DBUS_PROXY_FLAGS_NONE when creating the screencast portal DBus proxy.

Once we know the embedded cursor mode is available, add "cursor_mode"
parameter to the "SelectSources" method call. If that parameter is not
added, the hidden cursor mode is used by default.

Fixes: #238