nd-window: Add capsfilter for virtual display sources
For virtual displays, add a capsfilter named scrfilter in between the
pipewiresrc and the intervideosink.
With this filter, the source canvas resolution provided by mutter is
re-negotiated to the desired stream output resolution (currently hardcoded
to 1920x1080@30) instead of being lossily scaled to the desired resolution
from the default virtual display resolution hardcoded in mutter (1280x780@60).
To avoid confusion with the new capsfilter, the capsfilter previously named
sizefilter is renamed to sinkfilter.

Fixes #352