window: Re-create portal session when needed
Listen to the portal session's closed signal to stop the stream.
This signal is e.g. emitted when the "stop screencast" button on
GNOME shell's top bar is clicked.

Re-create the portal session if none exists when clicking on a
sink row. For now, one has to click on the sink row a second time
in order to start the actual stream.